% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{check_split}
\alias{check_split}
\title{Check if speciation happened}
\usage{
check_split(species_id, population, sex_dimensions, speciation_delta)
}
\arguments{
\item{species_id}{stub}

\item{population}{stub}

\item{sex_dimensions}{stub}

\item{speciation_delta}{stub}
}
\value{
Numeric with species ID of all individuals.
Otherwise returns \code{NULL}.
}
\description{
Check if speciation happened
}
