% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{competition_kernel}
\alias{competition_kernel}
\title{Calculate competition coefficient}
\usage{
competition_kernel(focal_id, competitor_id, niche_width, geo_width,
  eco_distance_matrix, geo_distance_matrix)
}
\arguments{
\item{focal_id}{stub}

\item{competitor_id}{stub}

\item{niche_width}{stub}

\item{geo_width}{stub}

\item{eco_distance_matrix}{stub}

\item{geo_distance_matrix}{stub}
}
\value{
numeric with competition coefficient
}
\description{
Calculate competition coefficient
}
