% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timoandraph.R
\name{do_things}
\alias{do_things}
\title{Stub function to temporarily wrap around loop. needs variables}
\usage{
do_things()
}
\value{
stub. simulation.
}
\description{
Stub function to temporarily wrap around loop. needs variables
}
\author{
Pedro Neves
}
