% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{find_mate}
\alias{find_mate}
\title{Mate finding}
\usage{
find_mate(individual_id, maxmatedistance, geo_distance_matrix, mateslope,
  mateintercept, population)
}
\arguments{
\item{individual_id}{stub}

\item{maxmatedistance}{stub}

\item{geo_distance_matrix}{stub}

\item{mateslope}{stub}

\item{mateintercept}{stub}

\item{population}{stub}
}
\value{
Numeric indicating mate id.
}
\description{
Mate finding
}
