% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Matrix addition}
\usage{
hello()
}
\arguments{
\item{L}{A list of matrices.}
}
\value{
The sum of all matrices.
}
\description{
This function calculates the sum of several matrices.
}
\author{
Raphael Scherrer
}
