% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{initialize_population}
\alias{initialize_population}
\title{Initialize the Population Data Frame}
\usage{
initialize_population(necol, nspatial, nmating, nindiv, bounds = c(0,
  10), sd = 1)
}
\arguments{
\item{necol}{stub}

\item{nspatial}{stub}

\item{nmating}{stub}

\item{nindiv}{stub}

\item{bounds}{stub}

\item{sd}{stub}
}
\value{
Data frame with population info.
}
\description{
Initialize the Population Data Frame
}
