% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m4_simul.R
\name{m4_simul}
\alias{m4_simul}
\title{Simulate the process}
\usage{
m4_simul(n_geo_dim, n_eco_dim)
}
\arguments{
\item{n_geo_dim}{number of geographical dimensions}

\item{n_eco_dim}{number of ecological dimensions}
}
\value{
stub
}
\description{
Simulate the process
}
