% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{make_distance_matrix}
\alias{make_distance_matrix}
\title{Distance matrix}
\usage{
make_distance_matrix(population, dimensions)
}
\arguments{
\item{population}{stub}

\item{dimensions}{stuv}
}
\value{
A distance matrix.
}
\description{
Distance matrix
}
