% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{mating_success}
\alias{mating_success}
\title{Check mating success}
\usage{
mating_success(mom_id, dad_id, sexslope, sexintercept)
}
\arguments{
\item{mom_id}{stub}

\item{dad_id}{stub}

\item{sexslope}{stub}

\item{sexintercept}{stub}
}
\value{
a numeric? int? probability? of mating happening
}
\description{
Check mating success
}
