% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{produce_offspring}
\alias{produce_offspring}
\title{Produce offspring}
\usage{
produce_offspring(mom_id, dad_id, dispersal_distance, eco_dimensions,
  sex_dimensions, geo_dimensions, mutation_rate_eco, mutation_rate_sex,
  mutational_effect_eco, mutational_effect_sex, bounds, individual_species)
}
\arguments{
\item{mom_id}{stub}

\item{dad_id}{stub}

\item{dispersal_distance}{stub}

\item{eco_dimensions}{stub}

\item{sex_dimensions}{stub}

\item{geo_dimensions}{stub}

\item{mutation_rate_eco}{stub}

\item{mutation_rate_sex}{stub}

\item{mutational_effect_eco}{stub}

\item{mutational_effect_sex}{stub}

\item{bounds}{stuv}

\item{individual_species}{stub}
}
\value{
Vector of stub.
}
\description{
Produce offspring
}
