% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{survive}
\alias{survive}
\title{Determines if individual survives}
\usage{
survive(individual_id, base_survival, eco_dimensions, eco_cutoff,
  geo_cutoff, eco_distance_matrix, geo_distance_matrix, niche_width,
  geo_width, resource_peaks, resource_width, max_carrying_capacity)
}
\arguments{
\item{individual_id}{stub}

\item{base_survival}{stub}

\item{eco_dimensions}{stub}

\item{eco_cutoff}{stub}

\item{geo_cutoff}{stub}

\item{eco_distance_matrix}{stub}

\item{geo_distance_matrix}{stub}

\item{niche_width}{stub}

\item{geo_width}{stub}

\item{resource_peaks}{stub}

\item{resource_width}{stub}

\item{max_carrying_capacity}{stub}
}
\value{
Integer 1 or 0 determining if individual survives (1) or dies (0).
}
\description{
Determines if individual survives
}
