% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{take_selfie}
\alias{take_selfie}
\title{Take a Snapshot of the Population}
\usage{
take_selfie(population, t)
}
\arguments{
\item{population}{stub}

\item{t}{stub}
}
\value{
Stub. csv file.
}
\description{
Take a Snapshot of the Population
}
