% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{update_register}
\alias{update_register}
\title{Update the phylogeny}
\usage{
update_register(register, population, t)
}
\arguments{
\item{register}{stub}

\item{population}{stub}

\item{t}{stub}
}
\value{
List with register of unique species?
}
\description{
Update the phylogeny
}
