% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{update_speciation}
\alias{update_speciation}
\title{Update the population with new species if speciation has happened}
\usage{
update_speciation(population, speciation_delta, minspecsize)
}
\arguments{
\item{population}{stub}

\item{speciation_delta}{stub}

\item{minspecsize}{stub}
}
\value{
?List with updated population
}
\description{
Update the population with new species if speciation has happened
}
