% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timoandraph.R
\name{do_things}
\alias{do_things}
\title{Stub function to temporarily wrap around loop. needs variables}
\usage{
do_things(necol, nspatial, nmating, nindiv, bounds, sd, tmax,
  maxmatedistance, mateslope, mateintercept, sexslope, sextintercept,
  dispersal_distance, mutation_rate_eco, mutation_rate_sex,
  mutational_effect_eco, mutational_effect_sex, base_survival, eco_cutoff,
  geo_cutoff, niche_width, geo_width, resource_width,
  max_carrying_capacity, specialization_delta, minspecsize, selfietime)
}
\arguments{
\item{necol}{stub}

\item{nspatial}{stub}

\item{nmating}{stub}

\item{nindiv}{stub}

\item{bounds}{stub}

\item{sd}{stub}

\item{tmax}{stub}

\item{maxmatedistance}{stub}

\item{mateslope}{stub}

\item{mateintercept}{stub}

\item{sexslope}{stub}

\item{sextintercept}{stub}

\item{dispersal_distance}{stub}

\item{mutation_rate_eco}{stub}

\item{mutation_rate_sex}{stub}

\item{mutational_effect_eco}{stub}

\item{mutational_effect_sex}{stub}

\item{base_survival}{stub}

\item{eco_cutoff}{stub}

\item{geo_cutoff}{stub}

\item{niche_width}{stub}

\item{geo_width}{stub}

\item{resource_width}{stub}

\item{max_carrying_capacity}{stub}

\item{specialization_delta}{stub}

\item{minspecsize}{stub}

\item{selfietime}{stub}
}
\value{
stub. simulation.
}
\description{
Stub function to temporarily wrap around loop. needs variables
}
\author{
Pedro Neves
}
