% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m4_simul.R
\name{m4_simul}
\alias{m4_simul}
\title{Simulate the process}
\usage{
m4_simul(nGeoDim, nEcoDim)
}
\arguments{
\item{nGeoDim}{number of geographical dimensions}

\item{nEcoDim}{number of ecological dimensions}
}
\value{
stub
}
\description{
Simulate the process
}
